Create performance tests for CPU implementations
This creates a very simple test harness for CPU implementations that
can be run in the browser. It is not dependent on any other libraries.
The harness supports a number of warmup runs and a number of timing
runs. It records the min, max, mean, and median times for the test
runs.

This works by creating a new entrypoint, `cpu_benchmark`, which is
loaded by the `test/perf/cpu_benchmark.html` file. You can fire up a
Node server to serve this by running:

```shell
node http-server ./
```

This change also adds the `bin2js` utility that can create JS files
out of binary ROM images.